UPLOAD/requirements.txt
flask==3.0.0
flask-cors==4.0.0 
Werkzeug==3.0.1
python-dotenv==1.0.0
ffmpeg-python==0.2.0 
uuid==1.30
psycopg2-binary==2.9.9
